python-influx-producer/requirements.txt
certifi==2022.6.15
influxdb-client==1.31.0
python-dateutil==2.8.2
python-dotenv==0.20.0
Rx==3.2.0
six==1.16.0
urllib3==1.26.11
